requirements.txt
flask>=2.2
pydantic>=1.10
pdfplumber>=0.10
python-docx>=1.1
langchain-deepseek>=0.1
langchain-core>=0.1
gunicorn>=21.2
